+++ CHANGES.rst
@@ -18,6 +18,7 @@ Unreleased
 * Add method :meth:`.Reddit.put` for HTTP PUT requests.
 * Add methods :meth:`.add_calendar` and :meth:`.add_community_list`.
 * Add methods :meth:`.add_image_widget` and :meth:`.upload_image`.
+* Add method :meth:`.add_custom_widget`.
 
 **Changed**